Updated wsjcpp_wsjcpp_arguments
